Doing some work

+++ Golf Game/src/Club.java
@@ -3,11 +3,11 @@ import java.util.ArrayList;
 public class Club
 	{
 		private String name;
-		private double distance;
+		private int distance;
 		private double accuracy;
 		private boolean favorite;
 		
-		public Club(String n, double d, double a, boolean f)
+		public Club(String n, int d, double a, boolean f)
 		{
 			name=n;
 			distance=d;
@@ -25,12 +25,12 @@ public class Club
 			this.name = name;
 		}
 
-		public double getDistance()
+		public int getDistance()
 			{
 				return distance;
 			}
 
-		public void setDistance(double distance)
+		public void setDistance(int distance)
 			{
 				this.distance = distance;
 			}
@@ -59,19 +59,19 @@ public class Club
 		
 		public static void addClub()
 		{
-			club.add(new Club("Sand", 80.0, 80.0, false));
-			club.add(new Club("Approach", 110.0, 78.0, false));
-			club.add(new Club("Pitch", 135.0, 76.0, false));
-			club.add(new Club("9 Iron", 150.0, 75.0, false));
-			club.add(new Club("8 Iron", 160.0, 74.0, false));
-			club.add(new Club("7 Iron", 170.0, 73.0, false));
-			club.add(new Club("6 Iron", 180.0, 70.0, false));
-			club.add(new Club("5 Iron", 190.0, 68.0, false));
-			club.add(new Club("4 Iron", 200.0, 65.0, false));
-			club.add(new Club("Hybrid", 210.0, 70.0, false));
-			club.add(new Club("5 Wood", 220.0, 62.0, false));
-			club.add(new Club("3 Wood", 240.0, 60.0, false));
-			club.add(new Club("Driver", 260.0, 50.0, false));
+			club.add(new Club("Sand", 80, 80.0, false));
+			club.add(new Club("Approach", 110, 78.0, false));
+			club.add(new Club("Pitch", 135, 76.0, false));
+			club.add(new Club("9 Iron", 150, 75.0, false));
+			club.add(new Club("8 Iron", 160, 74.0, false));
+			club.add(new Club("7 Iron", 170, 73.0, false));
+			club.add(new Club("6 Iron", 180, 70.0, false));
+			club.add(new Club("5 Iron", 190, 68.0, false));
+			club.add(new Club("4 Iron", 200, 65.0, false));
+			club.add(new Club("Hybrid", 210, 70.0, false));
+			club.add(new Club("5 Wood", 220, 62.0, false));
+			club.add(new Club("3 Wood", 240, 60.0, false));
+			club.add(new Club("Driver", 260, 50.0, false));
 			
 		
 		}
